requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database drivers (install based on your database)
psycopg2-binary==2.9.9  # For PostgreSQL
